refactor: product 패키지 생성
