+++ src/test/java/kitchenpos/domain/ProductTest.java
@@ -1,5 +1,6 @@
 package kitchenpos.domain;
 
+import kitchenpos.domain.product.Product;
 import org.junit.jupiter.api.Test;
 
 import java.math.BigDecimal;
